fix: move delete dashboard api to edit page
