+++ src/views/edit/deleteDashboardMutation.js
@@ -1,0 +1,5 @@
+export const deleteDashboardMutation = {
+    type: 'delete',
+    resource: 'dashboards',
+    id: ({ id }) => id,
+}
